requirements.txt
Django==3.0.7
celery==4.4.4
requests==2.23.0
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.2

# testing utilities
ipython==7.12.0
mixer==6.1.3
pytest==5.3.5
pytest-django==3.8.0
pytest-dotenv==0.4.0
pytest-env==0.6.2
pytest-mock==3.0.0
pytest-sugar==0.9.2
requests-mock==1.7.0
mock==4.0.2
